feat: separate config-step files
