+++ src/pages/Main/components/ConfigStep/ConfigStepContent.style.ts
